Begin writing release engineering chapter
I had a first version of it in my Google Doc, but this is only the first
part of it, and I'm extending it with more details about how to use Git.
On the one hand, I think this should be obvious to most parties. On the
other hand, from historical experience, it is not obvious to a scary
number of networks, so maybe it is good to elaborate a bit more.