Add custom anchors to build process page
I want to link to specific pages and sections when we give feedback to
networks, and for that it is useful when these links are stable. That
would not be the case if we use the autogenerated ones, because if we
rephrase them slightly, the anchor would change. Also, now we can write
a human summary as an "identifier" for the requirement, which may come
in handy in other places too.