Write about the "platform"
I sometimes get asked to explain what Chorus is or does, or to provide
an "architecture diagram". That doesn't really apply because our systems
and operations are so broad. We could provide an architecture diagram of
how we validate on a particular network, but for all of Chorus, I think
it would be just a big box with servers in it. Or maybe more useful is
our org chart, cf. Conway's law. But we definitely have infrastructure
that is worth describing somehow. Maybe it is pretentious to call it a
"platform", but it really is a platform. So let's add a chapter about
that.